COPILOT REFACTOR:
feat: refactor theme management by creating ThemeContext and useTheme hook; update ThemeProvider and ThemeToggle components
