+++ web/src/components/ThemeContext.jsx
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+
+export const ThemeProviderContext = createContext({
+  theme: 'system',
+  setTheme: () => null,
+});
