+++ web/src/components/ThemeToggle.jsx
@@ -1,6 +1,6 @@
 import { Moon, Sun } from 'lucide-react';
 import { Button } from './ui/button';
-import { useTheme } from './ThemeProvider';
+import { useTheme } from '../hooks/useTheme';
 
 export function ThemeToggle() {
   const { theme, setTheme } = useTheme();
